+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleasd
+
+- Always retry on connection issues
+
 ## v0.16.1
 - Add attachments to `sync-raw-email`